♻️ Extract commit_message.go
